Range Editor take in rangeList

+++ src/gamePlayer/SquareGameGrid.java
@@ -64,7 +64,7 @@ public class SquareGameGrid extends GameGrid{
     @Override
     protected void initializeGrid () {
 
-        myCurrentLocation = new Point2D.Double();
+        myCurrentLocation = new Point2D.Double(0,0);
 
         for(int i=0; i<r; i++){
             for(int j=0; j<c; j++){
